+++ src/main/kotlin/com/example/repository/UserRepository.kt
@@ -11,11 +11,11 @@ interface UserRepository {
     suspend fun insertUser(user: User): ServiceResult<User?>
 
     suspend fun getUsers(): ServiceResult<List<User?>>
-    suspend fun getUserById(userId: Int): ServiceResult<User?>
+    suspend fun getUserById(userId: Long): ServiceResult<User?>
     suspend fun getUsersByEmail(email: String): ServiceResult<List<User?>>
 
-    suspend fun updateUser(userId: Int, user: User): ServiceResult<User>
+    suspend fun updateUser(userId: Long, user: User): ServiceResult<User>
 
-    suspend fun deleteUser(userId: Int)
+    suspend fun deleteUser(userId: Long)
     suspend fun deleteUsers()
 }
